+++ src/graphics/shapes/circle.cpp
@@ -4,22 +4,22 @@
 namespace cvg
 {
 
-circle::circle(size_t x, size_t y, size_t r) :
+circle::circle(float x, float y, float r) :
   x_{x}, y_{y}, r_{r}
 {
 }
 
-size_t circle::x() const
+float circle::x() const
 {
   return x_;
 }
 
-size_t circle::y() const
+float circle::y() const
 {
   return y_;
 }
 
-size_t circle::r() const
+float circle::r() const
 {
   return r_;
 }
